refactor: unify navigation contracts and runtime boundaries

+++ src/planning/navigation_planning_backend/include/planner_core/log_utils.hpp
@@ -1,25 +1,8 @@
-/**
-* This file is part of SUPER
-*
-* Copyright 2025 Yunfan REN, MaRS Lab, University of Hong Kong, <mars.hku.hk>
-* Developed by Yunfan REN <renyf at connect dot hku dot hk>
-* for more information see <https://github.com/hku-mars/SUPER>.
-* If you use this code, please cite the respective publications as
-* listed on the above website.
-*
-* SUPER is free software: you can redistribute it and/or modify
-* it under the terms of the GNU Lesser General Public License as published by
-* the Free Software Foundation, either version 3 of the License, or
-* (at your option) any later version.
-*
-* SUPER is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-* GNU General Public License for more details.
-*
-* You should have received a copy of the GNU Lesser General Public License
-* along with SUPER. If not, see <http://www.gnu.org/licenses/>.
-*/
+/*
+ * Product-owned navigation implementation.
+ * Algorithmic provenance and external attributions are documented in the
+ * package documentation; they are not part of the runtime API or behaviour.
+ */
 
 #ifndef LOG_UTILS_HPP
 #define LOG_UTILS_HPP
@@ -243,7 +226,8 @@ namespace navigation_planning_backend {
 
         void replanBackupTrajectory(const BackupTrajOpt::Ptr &backup_traj_opt, Trajectory &traj) {
             if (exp_traj.empty() || backup_init_ps.size() == 0 || backup_init_t_vec.size() == 0
-                || (backup_init_ps.size()!=backup_init_t_vec.size())
+                || (backup_init_ps.size() !=
+                    static_cast<std::size_t>(backup_init_t_vec.size()))
             ) {
                 fmt::print(fg(fmt::color::indian_red),"Backup traj is empty\n");
                 return;
